Add Swift testing
