+++ Package.swift
@@ -1,7 +1,11 @@
 import PackageDescription
 
 let package = Package(
-  name: "HackersAtCambridgeWebsite",
+  name: "HaCWebsite",
+  targets: [
+    Target(name: "HaCWebsiteLib"),
+    Target(name: "HaCWebsite", dependencies: ["HaCWebsiteLib"])
+  ],
   dependencies: [
     .Package(url: "https://github.com/IBM-Swift/Kitura.git", majorVersion: 1, minor: 6),
     .Package(url: "https://github.com/IBM-Swift/HeliumLogger.git", majorVersion: 1, minor: 6),
