+++ Sources/HaCWebsite/main.swift
@@ -1,0 +1,3 @@
+import HaCWebsiteLib
+
+serveWebsite()
