+++ Sources/HaCWebsiteLib/events.swift
@@ -1,0 +1,3 @@
+func societyName () -> String {
+  return "Hackers at Cambridge"
+}
