+++ Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import HaCWebsiteLibTests
+
+XCTMain([
+     testCase(HaCWebsiteLibTests.allTests),
+])
