+++ package.json
@@ -4,7 +4,7 @@
   "description": "The official website of Cambridge's technology society",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "swift test",
     "gulp": "gulp"
   },
   "author": "",
